Adds internal error handling on custom validations
An unexpected error inside a custom validator, for example a failed db call, would get returned as a validation error, and that didn't seem very correct.
I implemented a way to distinguish internal errors that happen in a validator from validation errors, using the error message.
The handler function now recieves the validation errors and a value to indicate if an internal error happened, so it can forward the errors to the proper view.