Implements file uploading and file form
Adds a button on the nav bar to show the file form. This form accepts a file, a name and a folder.
When submitted, the fields get validated and if there are no errors, the file gets uploaded to the supabase bucket. If there are errors, the form gets rendered again, displaying the errors