Implements folder sharing
There is a new folder share button on the folder page, that displays a form to creates a new share instance with multiple share expiration time options.
This approach of creating a share record allows to have multiple and independent shares of the same folder.
When accessing a shared folder page, the page details are not present if the user is not logged in.

Also modified the app level error handling.