Implements folder delete
When deleting a folder, the folder itself and all of its files and shares are deleting (this is because the cascade option on the prisma schema). However, there isn't a way to ensure that all of the files on the supabase storage are deleted correctly or not deleted at all, or at least not a way that i know of after researching for some time.